feat(renamed): Pages to pages
